LayoutTests/fast/selectors/checked-direct-adjacent-style-update-3-expected.txt

Test direct adjacent style update on changes of the :checked state of input element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial state is not checked.
PASS getComputedStyle(document.getElementById("foo-with-renderer")).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.getElementById("bar-with-renderer")).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.getElementById("foo-without-renderer")).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.getElementById("bar-without-renderer")).backgroundColor is "rgb(255, 0, 0)"
Checking the boxes through API.
PASS getComputedStyle(document.getElementById("foo-with-renderer")).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.getElementById("bar-with-renderer")).backgroundColor is "rgb(0, 255, 0)"
PASS getComputedStyle(document.getElementById("foo-without-renderer")).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.getElementById("bar-without-renderer")).backgroundColor is "rgb(0, 255, 0)"
Unchecking through API
PASS getComputedStyle(document.getElementById("foo-with-renderer")).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.getElementById("bar-with-renderer")).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.getElementById("foo-without-renderer")).backgroundColor is "rgb(255, 255, 255)"
PASS getComputedStyle(document.getElementById("bar-without-renderer")).backgroundColor is "rgb(255, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
